+++ cfgrib/dataset.py
@@ -1,5 +1,5 @@
 #
-# Copyright 2017-2018 European Centre for Medium-Range Weather Forecasts (ECMWF).
+# Copyright 2017-2019 European Centre for Medium-Range Weather Forecasts (ECMWF).
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -31,8 +31,8 @@ import attr
 import numpy as np
 
 from . import __version__
+from . import bindings
 from . import cfmessage
-from . import eccodes
 from . import messages
 
 LOG = logging.getLogger(__name__)
@@ -108,18 +108,7 @@ ALL_HEADER_DIMS = ENSEMBLE_KEYS + VERTICAL_KEYS + DATA_TIME_KEYS + REF_TIME_KEYS
 ALL_KEYS = GLOBAL_ATTRIBUTES_KEYS + DATA_ATTRIBUTES_KEYS + GRID_TYPE_KEYS + ALL_HEADER_DIMS
 
 COORD_ATTRS = {
-    'time': {
-        'units': 'seconds since 1970-01-01T00:00:00+00:00', 'calendar': 'proleptic_gregorian',
-        'standard_name': 'forecast_reference_time', 'long_name': 'initial time of forecast',
-    },
-    'step': {
-        'units': 'hours',
-        'standard_name': 'forecast_period', 'long_name': 'time since forecast_reference_time',
-    },
-    'valid_time': {
-        'units': 'seconds since 1970-01-01T00:00:00+00:00', 'calendar': 'proleptic_gregorian',
-        'standard_name': 'time', 'long_name': 'time',
-    },
+    # geography
     'latitude': {
         'units': 'degrees_north',
         'standard_name': 'latitude', 'long_name': 'latitude',
@@ -128,9 +117,14 @@ COORD_ATTRS = {
         'units': 'degrees_east',
         'standard_name': 'longitude', 'long_name': 'longitude',
     },
-    'isobaricInhPa': {
-        'units': 'hPa', 'positive': 'down', 'stored_direction': 'decreasing',
-        'standard_name': 'air_pressure', 'long_name': 'pressure',
+    # vertical
+    'depthBelowLand': {
+        'units': 'm', 'positive': 'down', 'long_name': 'soil depth',
+        'standard_name': 'depth',
+    },
+    'depthBelowLandLayer': {
+        'units': 'm', 'positive': 'down', 'long_name': 'soil depth',
+        'standard_name': 'depth',
     },
     'hybrid': {
         'units': '1', 'positive': 'down', 'long_name': 'hybrid level',
@@ -140,10 +134,36 @@ COORD_ATTRS = {
         'units': 'm', 'positive': 'up', 'long_name': 'height above the surface',
         'standard_name': 'height',
     },
+    'isobaricInhPa': {
+        'units': 'hPa', 'positive': 'down', 'stored_direction': 'decreasing',
+        'standard_name': 'air_pressure', 'long_name': 'pressure',
+    },
+    'isobaricInPa': {
+        'units': 'Pa', 'positive': 'down', 'stored_direction': 'decreasing',
+        'standard_name': 'air_pressure', 'long_name': 'pressure',
+    },
+    'isobaricLayer': {
+        'units': 'Pa', 'positive': 'down',
+        'standard_name': 'air_pressure', 'long_name': 'pressure',
+    },
+    # ensemble
     'number': {
         'units': '1',
         'standard_name': 'realization', 'long_name': 'ensemble member numerical id',
-    }
+    },
+    # time
+    'step': {
+        'units': 'hours',
+        'standard_name': 'forecast_period', 'long_name': 'time since forecast_reference_time',
+    },
+    'time': {
+        'units': 'seconds since 1970-01-01T00:00:00', 'calendar': 'proleptic_gregorian',
+        'standard_name': 'forecast_reference_time', 'long_name': 'initial time of forecast',
+    },
+    'valid_time': {
+        'units': 'seconds since 1970-01-01T00:00:00', 'calendar': 'proleptic_gregorian',
+        'standard_name': 'time', 'long_name': 'time',
+    },
 }
 
 
@@ -217,7 +237,7 @@ class OnDiskArray(object):
             for header_indexes, offset in self.offsets.items():
                 # NOTE: fill a single field as found in the message
                 message = self.stream.message_from_file(file, offset=offset[0])
-                values = message.message_get('values', eccodes.CODES_TYPE_DOUBLE)
+                values = message.message_get('values', bindings.CODES_TYPE_DOUBLE)
                 array.__getitem__(header_indexes).flat[:] = values
         array[array == self.missing_value] = np.nan
         return array
@@ -239,7 +259,7 @@ class OnDiskArray(object):
                     continue
                 # NOTE: fill a single field as found in the message
                 message = self.stream.message_from_file(file, offset=offset[0])
-                values = message.message_get('values', eccodes.CODES_TYPE_DOUBLE)
+                values = message.message_get('values', bindings.CODES_TYPE_DOUBLE)
                 array_field.__getitem__(tuple(array_field_indexes)).flat[:] = values
 
         array = array_field[(Ellipsis,) + item[-self.geo_ndim:]]
@@ -259,6 +279,7 @@ GRID_TYPES_2D_NON_DIMENSION_COORDS = [
 def build_geography_coordinates(
         index,  # type: messages.FileIndex
         encode_cf,  # type: T.Sequence[str]
+        errors,  # type: str
         log=LOG,  # type: logging.Logger
 ):
     # type: (...) -> T.Tuple[T.Tuple[str, ...], T.Tuple[int, ...], T.Dict[str, Variable]]
@@ -291,7 +312,8 @@ def build_geography_coordinates(
                 attributes=COORD_ATTRS['longitude'],
             )
         except KeyError:  # pragma: no cover
-            log.warning('No latitudes/longitudes provided by ecCodes for gridType=%r', grid_type)
+            if errors != 'ignore':
+                log.warning('ecCodes provides no latitudes/longitudes for gridType=%r', grid_type)
     else:
         geo_dims = ('values',)
         geo_shape = (index.getone('numberOfPoints'),)
@@ -308,7 +330,8 @@ def build_geography_coordinates(
                 attributes=COORD_ATTRS['longitude'],
             )
         except KeyError:  # pragma: no cover
-            log.warning('No latitudes/longitudes provided by ecCodes for gridType=%r', grid_type)
+            if errors != 'ignore':
+                log.warning('ecCodes provides no latitudes/longitudes for gridType=%r', grid_type)
     return geo_dims, geo_shape, geo_coord_vars
 
 
@@ -332,7 +355,7 @@ def encode_cf_first(data_var_attrs, encode_cf=('parameter', 'time')):
     return coords_map
 
 
-def build_variable_components(index, encode_cf=(), filter_by_keys={}, log=LOG):
+def build_variable_components(index, encode_cf=(), filter_by_keys={}, log=LOG, errors='warn'):
     data_var_attrs_keys = DATA_ATTRIBUTES_KEYS[:]
     data_var_attrs_keys.extend(GRID_TYPE_MAP.get(index.getone('gridType'), []))
     data_var_attrs = enforce_unique_attributes(index, data_var_attrs_keys, filter_by_keys)
@@ -365,7 +388,7 @@ def build_variable_components(index, encode_cf=(), filter_by_keys={}, log=LOG):
     header_dimensions = tuple(d for d, c in coord_vars.items() if c.data.size > 1)
     header_shape = tuple(coord_vars[d].data.size for d in header_dimensions)
 
-    geo_dims, geo_shape, geo_coord_vars = build_geography_coordinates(index, encode_cf)
+    geo_dims, geo_shape, geo_coord_vars = build_geography_coordinates(index, encode_cf, errors)
     dimensions = header_dimensions + geo_dims
     shape = header_shape + geo_shape
     coord_vars.update(geo_coord_vars)
@@ -410,7 +433,7 @@ def dict_merge(master, update):
 
 
 def build_dataset_components(
-        stream, indexpath='{path}.{short_hash}.idx', filter_by_keys={}, errors='strict',
+        stream, indexpath='{path}.{short_hash}.idx', filter_by_keys={}, errors='warn',
         encode_cf=('parameter', 'time', 'geography', 'vertical'), timestamp=None, log=LOG,
 ):
     filter_by_keys = dict(filter_by_keys)
@@ -421,7 +444,7 @@ def build_dataset_components(
     for param_id, short_name, var_name in zip(param_ids, index['shortName'], index['cfVarName']):
         var_index = index.subindex(paramId=param_id)
         dims, data_var, coord_vars = build_variable_components(
-            var_index, encode_cf, filter_by_keys,
+            var_index, encode_cf, filter_by_keys, errors=errors,
         )
         if 'parameter' in encode_cf and var_name not in ('undef', 'unknown'):
             short_name = var_name
@@ -432,9 +455,11 @@ def build_dataset_components(
             dict_merge(variables, vars)
         except ValueError:
             if errors == 'ignore':
-                log.exception("skipping variable: paramId==%r shortName=%r", param_id, short_name)
-            else:
+                pass
+            elif errors == 'raise':
                 raise
+            else:
+                log.exception("skipping variable: paramId==%r shortName=%r", param_id, short_name)
     attributes = enforce_unique_attributes(index, GLOBAL_ATTRIBUTES_KEYS, filter_by_keys)
     encoding = {
         'source': stream.path,
@@ -446,7 +471,7 @@ def build_dataset_components(
     attributes_namespace = {
         'cfgrib_version': __version__,
         'cfgrib_open_kwargs': json.dumps(encoding),
-        'eccodes_version': eccodes.codes_get_api_version(),
+        'eccodes_version': bindings.codes_get_api_version(),
         'timestamp': timestamp or datetime.datetime.now().isoformat().partition('.')[0]
     }
     history_in = '{timestamp} GRIB to CDM+CF via ' \
@@ -466,7 +491,7 @@ class Dataset(object):
     encoding = attr.attrib(type=T.Dict[str, T.Any])
 
 
-def open_file(path, grib_errors='ignore', **kwargs):
+def open_file(path, grib_errors='warn', **kwargs):
     """Open a GRIB file as a ``cfgrib.Dataset``."""
     if 'mode' in kwargs:
         warnings.warn("the `mode` keyword argument is ignored and deprecated", FutureWarning)