+++ tests/test_20_messages.py
@@ -5,7 +5,7 @@ import os.path
 
 import pytest
 
-from cfgrib import eccodes
+from cfgrib import bindings
 from cfgrib import messages
 
 
@@ -39,7 +39,7 @@ def test_Message_read():
 
 
 def test_Message_write(tmpdir):
-    res = messages.Message.from_sample_name('regular_ll_pl_grib2', errors='strict')
+    res = messages.Message.from_sample_name('regular_ll_pl_grib2')
     assert res['gridType'] == 'regular_ll'
 
     res.message_set('Ni', 20)
@@ -54,6 +54,19 @@ def test_Message_write(tmpdir):
     res['pl'] = [2., 3.]
     assert res['pl'] == [2., 3.]
 
+    # warn on errors
+    res['centreDescription'] = 'DUMMY'
+    assert res['centreDescription'] != 'DUMMY'
+    res['edition'] = -1
+    assert res['edition'] != -1
+
+    # ignore errors
+    res.errors = 'ignore'
+    res['centreDescription'] = 'DUMMY'
+    assert res['centreDescription'] != 'DUMMY'
+
+    # raise errors
+    res.errors = 'raise'
     with pytest.raises(KeyError):
         res['centreDescription'] = 'DUMMY'
 
@@ -105,9 +118,9 @@ def test_make_message_schema():
 
     res = messages.make_message_schema(message, ['paramId', 'shortName', 'values', 'non-existent'])
 
-    assert res['paramId'] == (eccodes.CODES_TYPE_LONG, 1)
-    assert res['shortName'] == (eccodes.CODES_TYPE_STRING, 1, 256)
-    assert res['values'] == (eccodes.CODES_TYPE_DOUBLE, 7320)
+    assert res['paramId'] == (bindings.CODES_TYPE_LONG, 1)
+    assert res['shortName'] == (bindings.CODES_TYPE_STRING, 1, 256)
+    assert res['values'] == (bindings.CODES_TYPE_DOUBLE, 7320)
     assert res['non-existent'] == ()
 
 
@@ -168,11 +181,19 @@ def test_FileIndex_from_indexpath_or_filestream(tmpdir):
     )
     assert isinstance(res, messages.FileIndex)
 
-    # can't create nor read index file
+    # do not read nor create the index file
     res = messages.FileIndex.from_indexpath_or_filestream(
         messages.FileStream(str(grib_file)),
         ['paramId'],
         indexpath='',
+    )
+    assert isinstance(res, messages.FileIndex)
+
+    # can't create nor read index file
+    res = messages.FileIndex.from_indexpath_or_filestream(
+        messages.FileStream(str(grib_file)),
+        ['paramId'],
+        indexpath=str(tmpdir.join('non-existent-folder').join('non-existent-file')),
     )
     assert isinstance(res, messages.FileIndex)
 
@@ -211,4 +232,12 @@ def test_FileStream():
     # __file__ is not a GRIB, but contains the "GRIB" string, so it is a very tricky corner case
     res = messages.FileStream(str(__file__))
     with pytest.raises(EOFError):
+        res.first()
+
+    res = messages.FileStream(str(__file__), errors='ignore')
+    with pytest.raises(EOFError):
+        res.first()
+
+    res = messages.FileStream(str(__file__), errors='raise')
+    with pytest.raises(bindings.EcCodesError):
         res.first()